restart

+++ src/parse.c
@@ -9,7 +9,7 @@ static inline void body_List_free(List li)
     if (li.tag != tag_list)
         return ;
     
-    for (int i = 0; i < li.size; i++)
+    for (size_t i = 0; i < li.size; i++)
         List_free(li.list[i]);
 }
 
@@ -114,7 +114,7 @@ List escaping(const Strv str)
     assert(sb.capacity == 0 || sb.arr != NULL);
     // res.size = 0;
 
-    for (int i = 0; i < str.size; i++)
+    for (size_t i = 0; i < str.size; i++)
     {
         if (str.arr[i] == '\\')
         {
